% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioArgo.R
\name{read.bioArgo}
\alias{read.bioArgo}
\title{Read an MBARI BioArgo file}
\usage{
read.bioArgo(file, url = FALSE, cache = FALSE)
}
\arguments{
\item{file}{local filename or connection from which to read data. In the case where the argument \code{url=TRUE}, it can specify the float file without the \code{.TXT} extention.}

\item{url}{logical indicating if the file should be obtained directly via download from the MBARI website. Will also check to see if there is a locally cached copy in the current working directory (see \code{cache} argument).}

\item{cache}{ignored if \code{url=FALSE}. Logical indicating if the downloaded file should be cached locally.}
}
\value{
An \code{oce} \code{argo} object, containing the extra data columns
  corresponding to the biological data.
}
\description{
Reads a "Bio-Argo" formatted file into an \code{oce} \code{argo} object.
}
\details{
Reads the plain-text csv files obtained from \code{http://www.mbari.org/science/upper-ocean-systems/chemical-sensor-group/floatviz/}
}
\examples{
d <- read.bioArgo(system.file("extdata", "5145HawaiiQc.txt", package="bioArgo"))
ds <- as.section(d)
par(mfrow=c(2, 1))
plot(d)
plot(ds, xtype='time', which='oxygen')
}
\author{
Clark Richards, Dan Kelley, Cara Wilson
}
\seealso{
\code{read.argo}, \code{as.argo}, and \code{as.section} from the
  \code{oce} package
}

