% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbari.R
\name{read.argo.mbari}
\alias{read.argo.mbari}
\title{Read an MBARI BioArgo file}
\usage{
read.argo.mbari(file)
}
\arguments{
\item{file}{filename or connection from which to read data}
}
\value{
An \code{oce} \code{argo} object, containing the extra data columns
  corresponding to the biological data.
}
\description{
Reads a "Bio-Argo" formatted file into an \code{oce} \code{argo} object.
}
\details{
Reads the plain-text csv files obtained from \code{http://www.mbari.org/science/upper-ocean-systems/chemical-sensor-group/floatviz/}
}
\examples{
d <- read.argo.mbari(system.file("extdata", "5145HawaiiQc.txt", package="mbari"))
ds <- as.section(d)
par(mfrow=c(2, 1))
plot(d)
plot(ds, xtype='time', which='oxygen')
}
\author{
Clark Richards, Dan Kelley, Cara Wilson
}
\seealso{
\code{read.argo}, \code{as.argo}, and \code{as.section} from the
  \code{oce} package
}

